commit 2

+++ client/src/service/api.js
@@ -43,10 +43,6 @@ axiosInstance.interceptors.response.use(
   }
 );
 
-///////////////////////////////
-// If success -> returns { isSuccess: true, data: object }
-// If fail -> returns { isFailure: true, status: string, msg: string, code: int }
-//////////////////////////////
 const processResponse = (response) => {
   if (response?.status === 200) {
     return { isSuccess: true, data: response.data };
@@ -60,35 +56,33 @@ const processResponse = (response) => {
   }
 };
 
-///////////////////////////////
-// If success -> returns { isSuccess: true, data: object }
-// If fail -> returns { isError: true, status: string, msg: string, code: int }
-//////////////////////////////
 const ProcessError = async (error) => {
   if (error.response) {
-    // Request made and server responded with a status code
-    // that falls out of the range of 2xx
     if (error.response?.status === 403) {
-      // const { url, config } = error.response;
-      // console.log(error);
-      // try {
-      //     let response = await API.getRefreshToken({ token: getRefreshToken() });
-      //     if (response.isSuccess) {
-      sessionStorage.clear();
-      //         setAccessToken(response.data.accessToken);
+      try {
+        let response = await API.getRefreshToken({ token: getRefreshToken() });
+        if (response.isSuccess) {
+          sessionStorage.clear();
+          setAccessToken(response.data.accessToken);
 
-      //         const requestData = error.toJSON();
+          const requestData = error.toJSON();
 
-      //         let response1 = await axios({
-      //             method: requestData.config.method,
-      //             url: requestData.config.baseURL + requestData.config.url,
-      //             headers: { "content-type": "application/json", "authorization": getAccessToken() },
-      //             params: requestData.config.params
-      //         });
-      //     }
-      // } catch (error) {
-      //     return Promise.reject(error)
-      // }
+          let response1 = await axios({
+            method: requestData.config.method,
+            url: requestData.config.baseURL + requestData.config.url,
+            headers: {
+              "content-type": "application/json",
+              authorization: getAccessToken(),
+            },
+            params: requestData.config.params,
+          });
+
+          // Continue processing with the new access token
+          return processResponse(response1);
+        }
+      } catch (error) {
+        return Promise.reject(error);
+      }
     } else {
       console.log("ERROR IN RESPONSE: ", error.toJSON());
       return {
@@ -98,7 +92,6 @@ const ProcessError = async (error) => {
       };
     }
   } else if (error.request) {
-    // The request was made but no response was received
     console.log("ERROR IN RESPONSE: ", error.toJSON());
     return {
       isError: true,
@@ -106,7 +99,6 @@ const ProcessError = async (error) => {
       code: "",
     };
   } else {
-    // Something happened in setting up the request that triggered an Error
     console.log("ERROR IN RESPONSE: ", error.toJSON());
     return {
       isError: true,
